+++ solutions/geeksforgeeks/longest-common-substring/longest-common-substring.cpp
@@ -3,7 +3,7 @@
 #include <iostream>
 
 int main() {
-    int t, n, m, lcs;
+    int16_t t, n, m, lcs;
     std::string a, b;
 
     std::cin >> t;
@@ -16,12 +16,12 @@ int main() {
 
         lcs = 0;
 
-        int lcs_table[a.size() + 1][b.size() + 1];
+        int16_t lcs_table[a.size() + 1][b.size() + 1];
 
         memset(lcs_table, 0, sizeof(lcs_table));
 
-        for (int i = 1; i <= a.size(); i++) {
-            for (int j = 1; j <= b.size(); j++) {
+        for (int16_t i = 1; i <= a.size(); i++) {
+            for (int16_t j = 1; j <= b.size(); j++) {
                 if (a[i - 1] == b[j - 1]) {
                     lcs_table[i][j] = lcs_table[i - 1][j - 1] + 1;
                     lcs = std::max(lcs_table[i][j], lcs);
