+++ solutions/google-code-jam/2014/qualification-round/magic-trick/magic-trick.cpp
@@ -1,15 +1,16 @@
+#include <cstdint>
 #include <cstdio>
 
 int main() {
-    int cases, nrow, t, c, card, n = 0;
-    int first_row[4], second_row[4];
+    int16_t cases, nrow, t, c, card, n = 0;
+    int16_t first_row[4], second_row[4];
 
     scanf("%d", &cases);
 
     while (++n <= cases) {
         scanf("%d", &nrow);
-        for (int i = 1; i <= 4; ++i) {
-            for (int j = 0; j <= 3; ++j) {
+        for (int16_t i = 1; i <= 4; ++i) {
+            for (int16_t j = 0; j <= 3; ++j) {
                 scanf("%d", &t);
                 if (nrow == i) {
                     first_row[j] = t;
@@ -18,8 +19,8 @@ int main() {
         }
 
         scanf("%d", &nrow);
-        for (int i = 1; i <= 4; ++i) {
-            for (int j = 0; j <= 3; ++j) {
+        for (int16_t i = 1; i <= 4; ++i) {
+            for (int16_t j = 0; j <= 3; ++j) {
                 scanf("%d", &t);
                 if (nrow == i) {
                     second_row[j] = t;
@@ -28,8 +29,8 @@ int main() {
         }
 
         c = 0;
-        for (int i = 0; i <= 3; ++i) {
-            for (int j = 0; j <= 3; ++j) {
+        for (int16_t i = 0; i <= 3; ++i) {
+            for (int16_t j = 0; j <= 3; ++j) {
                 if (first_row[i] == second_row[j]) {
                     card = second_row[j];
                     c++;
