+++ solutions/google-code-jam/2016/qualification-round/counting-sheep/counting-sheep.cpp
@@ -1,9 +1,10 @@
+#include <cstdint>
 #include <iostream>
 #include <set>
 
 int main() {
-    int t, n, next, digits, i;
-    std::set< int > all_digits;
+    int16_t t, n, next, digits, i;
+    std::set< int16_t > all_digits;
 
     std::cin >> t;
