+++ solutions/project-euler/021/021.cpp
@@ -1,8 +1,8 @@
 #include <iostream>
 #include <set>
 
-int sum_divisors(int n) {
-    int sum = 1, i;
+int16_t sum_divisors(int16_t n) {
+    int16_t sum = 1, i;
 
     for (i = 2; i <= n / 2; i++) {
         if (n % i == 0) {
@@ -13,14 +13,14 @@ int sum_divisors(int n) {
     return sum;
 }
 
-bool is_amicable(int n) {
+bool is_amicable(int16_t n) {
     return n == sum_divisors(sum_divisors(n));
 }
 
 int main() {
-    int i, sum = 0;
-    std::set< int > amicables;
-    std::set< int >::iterator it;
+    int16_t i, sum = 0;
+    std::set< int16_t > amicables;
+    std::set< int16_t >::iterator it;
 
     for (i = 1; i < 10000; i++) {
         if (is_amicable(i) && i != sum_divisors(i)) {
