+++ solutions/spoj-br/quadrad2/quadrad2.cpp
@@ -1,8 +1,12 @@
+#include <cstdint>
 #include <cstdio>
 
 int main() {
-    int n;
+    int16_t n;
+
     scanf("%d", &n);
+
     printf("%d\n", n * n);
+
     return 0;
 }
