+++ solutions/spoj-br/soma/soma.cpp
@@ -1,12 +1,17 @@
+#include <cstdint>
 #include <cstdio>
 
 int main() {
-    int i, n, s = 0;
+    int16_t i, n, s = 0;
+
     scanf("%d", &i);
+
     while (i--) {
         scanf("%d", &n);
         s += n;
     }
+
     printf("%d\n", s);
+
     return 0;
 }
