+++ solutions/uri/1007/1007.cpp
@@ -1,7 +1,8 @@
+#include <cstdint>
 #include <cstdio>
 
 int main() {
-    int a, b, c, d;
+    int16_t a, b, c, d;
 
     scanf("%d", &a);
     scanf("%d", &b);
