+++ solutions/uri/1013/1013.cpp
@@ -2,7 +2,7 @@
 #include <cstdio>
 
 int main() {
-    int a, b, c, m;
+    int16_t a, b, c, m;
 
     while (scanf("%d %d %d", &a, &b, &c) != EOF) {
         m = (a + b + std::abs(a - b)) / 2;
