+++ solutions/uri/1024/1024.cpp
@@ -2,7 +2,7 @@
 #include <string>
 
 int main() {
-    int n;
+    int16_t n;
     std::string s;
 
     std::cin >> n;
@@ -13,9 +13,9 @@ int main() {
 
         std::string r(s.rbegin(), s.rend());  // reverse string
 
-        unsigned int l = r.length();
+        uint16_t l = r.length();
 
-        for (unsigned int i = 0; i < s.length(); ++i) {
+        for (uint16_t i = 0; i < s.length(); ++i) {
             if ((r.at(i) >= 65 && r.at(i) <= 90)
                 || (r.at(i) >= 97 && r.at(i) <= 122)) {
                 r.at(i) = r.at(i) + 3;
