+++ solutions/uri/1029/1029.cpp
@@ -1,10 +1,11 @@
+#include <cstdint>
 #include <cstdio>
 #include <cstring>
 
-long int f[39];
-long int r[39];
+int32_t f[39];
+int32_t r[39];
 
-long int fib(long int n) {
+int32_t fib(int32_t n) {
     if (n == 0) {
         return f[0];
     }
@@ -20,8 +21,8 @@ long int fib(long int n) {
 }
 
 int main() {
-    int i, j;
-    long int n;
+    int16_t i, j;
+    int32_t n;
 
     memset(f, 0, sizeof(f));
     memset(r, 0, sizeof(r));
