+++ solutions/uri/1059/1059.cpp
@@ -1,7 +1,8 @@
+#include <cstdint>
 #include <iostream>
 
 int main() {
-    for (int i = 2; i <= 100; i += 2) {
+    for (int16_t i = 2; i <= 100; i += 2) {
         std::cout << i << std::endl;
     }
     return 0;
