+++ solutions/uri/1060/1060.cpp
@@ -1,13 +1,17 @@
+#include <cstdint>
 #include <cstdio>
 
 int main() {
     float n;
-    int c = 0;
+    int16_t c = 0;
+
     while (scanf("%f", &n) != EOF) {
         if (n > 0) {
             c++;
         }
     }
+
     printf("%d valores positivos\n", c);
+
     return 0;
 }
