+++ solutions/uri/1064/1064.cpp
@@ -1,9 +1,10 @@
 #include <cmath>
+#include <cstdint>
 #include <cstdio>
 
 int main() {
     float n, s = 0;
-    int c = 0;
+    int16_t c = 0;
 
     while (scanf("%f", &n) != EOF) {
         if (n > 0) {
