+++ solutions/uri/1066/1066.cpp
@@ -1,7 +1,8 @@
+#include <cstdint>
 #include <cstdio>
 
 int main() {
-    int c = 5, x, p = 0, n = 0, e = 0, o = 0;
+    int16_t c = 5, x, p = 0, n = 0, e = 0, o = 0;
 
     while (c--) {
         scanf("%d", &x);
