+++ solutions/uri/1067/1067.cpp
@@ -1,10 +1,14 @@
+#include <cstdint>
 #include <iostream>
 
 int main() {
-    int n, i;
+    int16_t n, i;
+
     std::cin >> n;
+
     for (i = 1; i <= n; i += 2) {
         std::cout << i << std::endl;
     }
+
     return 0;
 }
