+++ solutions/uri/1069/1069.cpp
@@ -1,10 +1,11 @@
+#include <cstdint>
 #include <cstdio>
 #include <cstring>
 #include <iostream>
 
 int main() {
     char e[1001];
-    int d, n, c, i;
+    int16_t d, n, c, i;
 
     std::cin >> n;
     getchar();  // pra obrigar o cursor ir para a linha de baixo
@@ -15,7 +16,7 @@ int main() {
 
         fgets(e, 1000, stdin);
 
-        for (d = 0; d < static_cast< int >(strlen(e)); d++) {
+        for (d = 0; d < static_cast< int16_t >(strlen(e)); d++) {
             if (e[d] == '<') {
                 i++;
             } else if (e[d] == '>' && i - 1 >= 0) {
