+++ solutions/uri/1070/1070.cpp
@@ -1,14 +1,15 @@
+#include <cstdint>
 #include <iostream>
 
 int main() {
-    int n;
+    int16_t n;
 
     std::cin >> n;
 
     if (n % 2 == 0) {
         n++;
     }
-    for (int i = 0; i <= 5; i++, n += 2) {
+    for (int16_t i = 0; i <= 5; i++, n += 2) {
         std::cout << n << std::endl;
     }
