+++ solutions/uri/1080/1080.cpp
@@ -1,7 +1,8 @@
+#include <cstdint>
 #include <cstdio>
 
 int main() {
-    int n, q = 0, g = -1, i = -1;
+    int16_t n, q = 0, g = -1, i = -1;
 
     while (++q && q <= 100) {
         scanf("%d", &n);
